Delete showCustomKeyboardAler method from BaseActivity
