+++ app/src/main/java/io/horizontalsystems/bankwallet/core/BaseActivity.kt
@@ -1,46 +1,14 @@
 package io.horizontalsystems.bankwallet.core
 
-import android.content.Context
 import android.os.Bundle
-import android.os.Handler
 import android.view.View
 import android.view.inputmethod.InputMethodManager
 import androidx.navigation.NavController
 import androidx.navigation.NavDestination
-import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.core.CoreActivity
 import io.horizontalsystems.core.hideKeyboard
-import io.horizontalsystems.views.AlertDialogKeyboardFragment
-
 
 abstract class BaseActivity : CoreActivity(), NavController.OnDestinationChangedListener {
-
-    fun showCustomKeyboardAlert() {
-        AlertDialogKeyboardFragment.newInstance(
-                titleString = getString(R.string.Alert_TitleWarning),
-                descriptionString = getString(R.string.Alert_CustomKeyboardIsUsed),
-                selectButtonText = R.string.Alert_Select,
-                skipButtonText = R.string.Alert_Skip,
-                listener = object : AlertDialogKeyboardFragment.Listener {
-                    override fun onButtonClick() {
-                        val imeManager = getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
-                        imeManager.showInputMethodPicker()
-                        hideSoftKeyboard()
-                        Handler().postDelayed({
-                            try {
-                                onBackPressed()
-                            } catch (e: NullPointerException) {
-                                //do nothing
-                            }
-                        }, (1 * 750).toLong())
-                    }
-
-                    override fun onCancel() {}
-                    override fun onSkipClick() {
-                        App.thirdKeyboardStorage.isThirdPartyKeyboardAllowed = true
-                    }
-                }).show(supportFragmentManager, "custom_keyboard_alert")
-    }
 
     protected fun setTransparentStatusBar() {
         val oldFlags = window.decorView.systemUiVisibility
